[#5] FCM 연결
테스트 완료
