+++ Hyodor/Hyodor/GoogleService-Info.plist
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>API_KEY</key>
+	<string>AIzaSyDoTV-bSh8JaXluYQ9eXKsp6hWN5PbBdIg</string>
+	<key>GCM_SENDER_ID</key>
+	<string>491051779559</string>
+	<key>PLIST_VERSION</key>
+	<string>1</string>
+	<key>BUNDLE_ID</key>
+	<string>com.SangjunKim.Hyodor</string>
+	<key>PROJECT_ID</key>
+	<string>hydor-6f2b8</string>
+	<key>STORAGE_BUCKET</key>
+	<string>hydor-6f2b8.firebasestorage.app</string>
+	<key>IS_ADS_ENABLED</key>
+	<false></false>
+	<key>IS_ANALYTICS_ENABLED</key>
+	<false></false>
+	<key>IS_APPINVITE_ENABLED</key>
+	<true></true>
+	<key>IS_GCM_ENABLED</key>
+	<true></true>
+	<key>IS_SIGNIN_ENABLED</key>
+	<true></true>
+	<key>GOOGLE_APP_ID</key>
+	<string>1:491051779559:ios:aba11d22634e73249c2942</string>
+</dict>
+</plist>
